Bug Fix for Rotation after zoom transition. New FlickrNetworkManager vs FlickrDataManager
